Deploy/requirements.txt
Flask==3.0.3
torch==2.5.1
torchvision==0.20.1
Pillow==10.4.0
Werkzeug==3.0.4